requirements.txt
requests==2.31.0
transformers==4.44.2
diffusers==0.30.3
Pillow==10.4.0
googlemaps==4.10.0
sentencepiece==0.2.0
protobuf==5.28.2
accelerate==0.34.2
bitsandbytes==0.43.3
flash_attn==2.6.3
huggingface_hub==0.25.1
configargparse==1.7